19 nov  start

+++ services/isOtpValid.js
@@ -1,19 +1,17 @@
 import OTP from "../model/OTP.js";
 
-export const isOtpValid = async (email, otp) => {
+export const isOtpValid = async (email, otp,isDeleteOtpDocument=true) => {
   try {
     const otpObj = await OTP.findOne({ email, otp });
+    console.log({otpObj});
     //now i am deleting otp for all time. but after some time i will take an argument
-    if(otpObj){
-      const a =  await otpObj.deleteOne()
-      console.log({a});
-        
+    if (otpObj && isDeleteOtpDocument) {
+      const a = await otpObj.deleteOne();
+      console.log({ a });
     }
-    return otpObj ? true : false
-
+    return otpObj ? true : false;
   } catch (error) {
-
     console.log(error);
-    return false
+    return false;
   }
 };
